requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
bcrypt
python-jose